devide all to files
